CHANGE: Minimum load is now stored as a uint16 value, allowing minimum load ratios of over 255 (current maximum is set to 400) CHANGE: Minimum load increment is now 10% rather than an automatic progression ***WARNING***: Games saved with previous non-release compiles - including the release candidate binary - will not load with this version or the next official release because of the above changes. Save your games with a lower version before updating.

+++ dataobj/fahrplan.h
@@ -119,11 +119,11 @@ public:
 	 * fügt eine koordinate an stelle aktuell in den Fahrplan ein
 	 * alle folgenden Koordinaten verschieben sich dadurch
 	 */
-	bool insert(const grund_t* gr, uint8 ladegrad = 0, uint8 waiting_time_shift = 0,  sint16 spacing_shift = 0, bool show_failure = false);
+	bool insert(const grund_t* gr, uint16 ladegrad = 0, uint8 waiting_time_shift = 0,  sint16 spacing_shift = 0, bool show_failure = false);
 	/**
 	 * hängt eine koordinate an den fahrplan an
 	 */
-	bool append(const grund_t* gr, uint8 ladegrad = 0, uint8 waiting_time_shift = 0, sint16 spacing_shift = 0);
+	bool append(const grund_t* gr, uint16 ladegrad = 0, uint8 waiting_time_shift = 0, sint16 spacing_shift = 0);
 
 	// cleanup a schedule, removes double entries
 	void cleanup();
